refactor: additional test case
